refatorando  a partir dos comentários propostos, metodo isService colocado no controller de aluno
