+++ aplicacoes/gpa-mae/src/main/java/br/ufc/quixada/npi/gpa/service/impl/AlunoServiceImpl.java
@@ -8,8 +8,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import br.ufc.quixada.npi.enumeration.QueryType;
 import br.ufc.quixada.npi.gpa.model.Aluno;
-import br.ufc.quixada.npi.gpa.model.Inscricao;
-import br.ufc.quixada.npi.gpa.model.Selecao;
 import br.ufc.quixada.npi.gpa.service.AlunoService;
 import br.ufc.quixada.npi.service.impl.GenericServiceImpl;
 import br.ufc.quixada.npi.util.SimpleMap;
@@ -49,13 +47,6 @@ public class AlunoServiceImpl extends GenericServiceImpl<Aluno> implements Aluno
 		return (Aluno) findFirst("Aluno.findAlunoByCPF", new SimpleMap<String, Object>("cpf", cpf));
 	}
 
-	@Override
-	public boolean isAlunoCadastradoEmSelecao(Aluno aluno, Selecao selecao) {
-		for (Inscricao inscricao : aluno.getInscricoes()) {
-			if(inscricao.getSelecao().equals(selecao))
-				return true;
-		}
-		return false;
-	}
+	
 
 }
